# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/publisher": {
            "get": {
                "description": "Get a list of Publisher.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Get all Publisher.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Publisher"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new Publisher.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Create New Publisher.",
                "parameters": [
                    {
                        "description": "the body to create a new Publisher",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.publisherInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publisher"
                        }
                    }
                }
            }
        },
        "/publisher/{id}": {
            "get": {
                "description": "Get an Publisher by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Get Publisher.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Publisher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publisher"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a Publisher by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Delete one Publisher.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Publisher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update Publisher by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Update Publisher.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Publisher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update Publisher",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.publisherInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Publisher"
                        }
                    }
                }
            }
        },
        "/publisher/{id}/game": {
            "get": {
                "description": "Get all Game by PublisherId.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Publisher"
                ],
                "summary": "Get Game.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Publisher id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Game"
                            }
                        }
                    }
                }
            }
        },
        "/ratinggame": {
            "get": {
                "description": "Get a list of RatingGame.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RatingGame"
                ],
                "summary": "Get all RatingGame.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Rating"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Creating a new RatingGame.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RatingGame"
                ],
                "summary": "Create New RatingGame.",
                "parameters": [
                    {
                        "description": "the body to create a new game",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ratingInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Game"
                        }
                    }
                }
            }
        },
        "/ratinggame/{id}": {
            "get": {
                "description": "Get a RatingGame by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RatingGame"
                ],
                "summary": "Get RatingGame.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ratinggame id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Rating"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Delete a RatingGame by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RatingGame"
                ],
                "summary": "Delete one RatingGame.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ratinggame id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerToken": []
                    }
                ],
                "description": "Update RatingGame by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RatingGame"
                ],
                "summary": "Update RatingGame.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ratinggame id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update an RatingGame",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ratingInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization. How to input in swagger : 'Bearer \u003cinsert_your_token_here\u003e'",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Rating"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.publisherInput": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "controller.ratingInput": {
            "type": "object",
            "properties": {
                "game_id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                }
            }
        },
...
        "models.Game": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "age_rating_category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "developer_id": {
                    "type": "integer"
                },
                "genre_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "publisher_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "models.Publisher": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Rating": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "gameID": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}